requirements.txt
tensorflow-gpu==2.6.0
keras==2.6.0
numpy==1.19.5

scipy==1.10.1
matplotlib==3.3.4
scikit-image==0.18.2
opencv-python==4.10.0.84

h5py==3.1.0
ipython==8.12.2
pillow==10.4.0
pycocotools==2.0.7
PyYAML==6.0.1
imgaug==0.4.0
